# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Eligo Backend API
  version: 1.7.0
  description: |
    Eligo 后端接口的唯一机器可读契约。
    M1 个人资料契约采用“同一账号叠加能力”模型：普通参与和个人发布不拆分账号身份，
    用户与后端预置企业建立有效 OWNER 关系后，再叠加企业操作能力。
    当前不提供企业申请、认证、审核或身份切换接口。
    新增或修改接口时，后端开发者必须在同一个提交中更新本文件。
    除文件内容下载接口外，所有 JSON 响应都使用统一 Result 包装。
  x-generated-from-code-date: 2026-08-22
  x-contract-updated-date: 2026-08-22
  x-m1-contract-status: REVIEW
  x-m2-contract-status: IMPLEMENTED
  x-m3-contract-status: IMPLEMENTED
  x-m4-contract-status: IMPLEMENTED
  x-m4.1-contract-status: IMPLEMENTED
servers:
  - url: http://127.0.0.1:8080
    description: 本机 Docker / 微信开发者工具
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/users/me/posts:
    post:
      tags: [Post]
      operationId: createMyPostDraft
      summary: 创建个人动态草稿
      description: |
        要求当前用户具备 PUBLISH_PERSONAL_POST；资料未完成返回 11201/409。
        只创建 DRAFT，不直接发布。标题、正文、图片或活动引用至少存在一项；默认可见范围为 PUBLIC。
        POST 图片最多 5 张，只接受当前用户上传且已通过内容检查的 JPEG 或 PNG；
        单文件不超过 10 MiB，总计不超过 50 MiB。活动引用只接受当前公开可读的
        PUBLISHED、CANCELLED 或 ENDED 活动，不可用时返回 11906/409。
        幂等结果从首次创建起固定保留 24 小时：同键同请求重放返回原草稿与 200，
        同键不同请求返回 11901/409；原草稿已软删除时返回 11904/409，重放不延长窗口。
      x-error-precedence:
        - PROFILE_AND_CAPABILITY_AUTHORIZATION
        - IDEMPOTENCY_REPLAY_OR_CONFLICT
        - CONTENT_AND_REFERENCE_VALIDATION
        - CREATE_DRAFT
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostDraftCreateRequest"
      responses:
        "200":
          description: 24 小时内同键同请求的幂等重放结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "201":
          description: 个人动态草稿已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/organizations/{organizationId}/posts:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    post:
      tags: [Post]
      operationId: createOrganizationPostDraft
      summary: 创建企业动态草稿
      description: |
        当前用户资料未完成时返回 11201/409；通过资料门槛后再校验企业写权限。
        企业必须处于 DEV_ENABLED；当前用户必须具备 PUBLISH_ORGANIZATION_POST，且是当前有效且唯一的 OWNER。
        草稿归企业主体，不归操作人个人；
        只创建 DRAFT，不直接发布。字段、图片、活动引用和 24 小时幂等规则与个人草稿一致：
        同键同请求返回原草稿与 200，同键不同请求返回 11901/409，原草稿已软删除返回 11904/409。
      x-error-precedence:
        - PROFILE_COMPLETENESS_11201
        - ORGANIZATION_WRITE_AUTHORIZATION_404
        - IDEMPOTENCY_REPLAY_OR_CONFLICT
        - CONTENT_AND_REFERENCE_VALIDATION
        - CREATE_DRAFT
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostDraftCreateRequest"
      responses:
        "200":
          description: 24 小时内同键同请求的幂等重放结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "201":
          description: 企业动态草稿已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
    get:
      tags: [Post]
      operationId: listOrganizationPosts
      summary: 游标分页查询企业动态
      description: |
        固定按 publishedAt DESC、postId DESC 排序。匿名只返回 PUBLISHED + PUBLIC；
        携带登录令牌时，在读取时实时校验关注关系和企业 OWNER 权限后，可补充 FOLLOWERS_ONLY 或 PRIVATE。
        企业不存在或企业不可用时统一返回 404；历史动态保留在数据库中，但不再公开分发。
      x-implementation-status: IMPLEMENTED
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FilterBoundCursor"
        - $ref: "#/components/parameters/PostLimit"
      responses:
        "200":
          description: 企业动态分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostPageResult"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/me/posts/{postId}:
    parameters:
      - $ref: "#/components/parameters/PostId"
    put:
      tags: [Post]
      operationId: replaceMyPostDraft
      summary: 完整替换个人动态草稿
      description: |
        要求当前用户具备 PUBLISH_PERSONAL_POST；只有当前用户的 DRAFT 可编辑。
        必须提供当前 version，版本不一致返回 11903/409；资料未完成返回 11201/409。
        保存时重新校验活动引用，只接受当前公开可读的 PUBLISHED、CANCELLED 或 ENDED 活动；
        不可用时返回 11906/409。草稿保存事务绑定并激活新图片，旧的无引用图片进入清理流程。
      x-error-precedence:
        - RESOURCE_MANAGE_VISIBILITY_404
        - PROFILE_AND_CAPABILITY_AUTHORIZATION
        - VERSION_11903
        - STATUS_11902
        - CONTENT_AND_REFERENCE_VALIDATION
        - REPLACE_DRAFT
      x-implementation-status: IMPLEMENTED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostDraftReplaceRequest"
      responses:
        "200":
          description: 更新后的动态草稿
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/organizations/{organizationId}/posts/{postId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/PostId"
    put:
      tags: [Post]
      operationId: replaceOrganizationPostDraft
      summary: 完整替换企业动态草稿
      description: |
        当前用户资料未完成时返回 11201/409；通过资料门槛后再校验企业写权限。
        企业必须处于 DEV_ENABLED；当前用户必须具备 PUBLISH_ORGANIZATION_POST，且是当前有效且唯一的 OWNER，
        才可编辑该企业 DRAFT。版本不一致返回 11903/409；无管理权限统一返回 404。
        活动引用只接受当前公开可读的 PUBLISHED、CANCELLED 或 ENDED 活动，不可用时返回 11906/409；
        草稿保存事务绑定并激活新图片，旧的无引用图片进入清理流程。
      x-error-precedence:
        - RESOURCE_MANAGE_VISIBILITY_404
        - PROFILE_COMPLETENESS_11201
        - ORGANIZATION_WRITE_AUTHORIZATION_404
        - VERSION_11903
        - STATUS_11902
        - CONTENT_AND_REFERENCE_VALIDATION
        - REPLACE_DRAFT
      x-implementation-status: IMPLEMENTED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostDraftReplaceRequest"
      responses:
        "200":
          description: 更新后的企业动态草稿
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/posts/{postId}/publication:
    parameters:
      - $ref: "#/components/parameters/PostId"
    put:
      tags: [Post]
      operationId: publishPost
      summary: 发布动态草稿
      description: |
        仅允许 DRAFT 转为 PUBLISHED：发布要求标题有效，且正文或图片至少存在一项；
        重新校验资料完成度、活动引用和全部图片。已经 PUBLISHED 时幂等返回当前结果。
        HIDDEN 是平台内容安全状态，作者不得自行恢复，调用本接口返回 11902/409；
        HIDDEN 转回 PUBLISHED 仅允许未公开的内部平台审核或申诉流程执行，并保留原 publishedAt。
        个人动态要求 PUBLISH_PERSONAL_POST；企业动态要求企业处于 DEV_ENABLED、当前用户具备
        PUBLISH_ORGANIZATION_POST，且是当前有效且唯一的 OWNER；无管理权限统一返回 404。
        判定时先确认当前资源管理权限；已经 PUBLISHED 时优先幂等返回，不因资料或企业能力后来变化而重复写入；
        首次发布才重新校验写能力、内容、活动引用和图片。
      x-error-precedence:
        - RESOURCE_MANAGE_VISIBILITY_404
        - PUBLISHED_IDEMPOTENCY
        - WRITE_AUTHORIZATION
        - STATUS_11902
        - CONTENT_AND_REFERENCE_VALIDATION
        - PUBLISH_POST
      x-implementation-status: IMPLEMENTED
      responses:
        "200":
          description: 已发布动态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/posts/{postId}:
    parameters:
      - $ref: "#/components/parameters/PostId"
    get:
      tags: [Post]
      operationId: getPost
      summary: 获取动态详情
      description: |
        普通详情只返回 PUBLISHED；DRAFT、HIDDEN、DELETED 或不存在统一返回 404。
        允许可选登录。PUBLIC 可匿名读取；FOLLOWERS_ONLY 在读取时实时检查当前关注关系，
        个人作者本人或企业当前有效且唯一的 OWNER 也可读取自己主体的 FOLLOWERS_ONLY；
        PRIVATE 个人动态只允许个人作者本人，企业动态只允许企业当前有效且唯一的 OWNER。
        作者账号或企业不可用时，普通读取与公开分发统一返回 404；管理者查看历史须使用可管理动态接口。
        其他无权读取场景统一返回 404。
      x-implementation-status: IMPLEMENTED
      security:
        - {}
        - bearerAuth: []
      responses:
        "200":
          description: 动态详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostResult"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags: [Post]
      operationId: deletePost
      summary: 软删除本人可管理的动态
      description: |
        DRAFT、PUBLISHED 或 HIDDEN 均可转为 DELETED；重复删除按资源不可见返回 404。
        个人动态只校验已认证登录的作者本人，不依赖资料完成度或 PUBLISH_PERSONAL_POST；
        企业动态要求企业处于 DEV_ENABLED、
        当前用户具备 PUBLISH_ORGANIZATION_POST，且是企业当前有效且唯一的 OWNER。
      x-error-precedence:
        - RESOURCE_MANAGE_VISIBILITY_404
        - DELETED_404
        - ORGANIZATION_WRITE_AUTHORIZATION_404
        - DELETE_POST
      x-implementation-status: IMPLEMENTED
      responses:
        "200":
          $ref: "#/components/responses/VoidSuccess"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/me/managed-posts:
    get:
      tags: [Post]
      operationId: listMyManagedPosts
      summary: 游标分页查询本人可管理的动态
      description: |
        返回本人的个人动态，以及当前作为有效且唯一 OWNER 可管理的企业动态。
        当前 OWNER 在企业变为 DEV_DISABLED 后仍可读取企业草稿、私密和隐藏历史，但不可写；
        OWNER 关系失效后不再返回该企业的非公开动态。
        固定按 updatedAt DESC、postId DESC 排序；不返回已删除动态。
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/FilterBoundCursor"
        - $ref: "#/components/parameters/PostLimit"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ManagedPostStatus"
        - name: authorType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PostAuthorType"
      responses:
        "200":
          description: 可管理动态分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostPageResult"
        "401":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/me/managed-posts/{postId}:
    parameters:
      - $ref: "#/components/parameters/PostId"
    get:
      tags: [Post]
      operationId: getMyManagedPost
      summary: 获取本人可管理的动态详情
      description: |
        个人作者或企业当前有效且唯一的 OWNER 可读取草稿、私密和隐藏内容；无权限统一返回 404。
        当前 OWNER 在企业变为 DEV_DISABLED 后仍可读取历史，但不可写；OWNER 关系失效后不再可读。
      x-implementation-status: IMPLEMENTED
      responses:
        "200":
          description: 可管理动态详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedPostDetailResult"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/posts:
    get:
      tags: [Post]
      operationId: listPublicPosts
      summary: 游标分页查询公开动态
      description: |
        只返回 PUBLISHED + PUBLIC，且用户或企业作者当前公开可用的动态。
        作者不可用时历史动态仍保留在数据库中但不公开分发；固定按 publishedAt DESC、postId DESC 排序，不包含推荐排序。
      x-implementation-status: IMPLEMENTED
      security: []
      parameters:
        - $ref: "#/components/parameters/FilterBoundCursor"
        - $ref: "#/components/parameters/PostLimit"
      responses:
        "200":
          description: 公开动态分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostPageResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/users/{userId}/posts:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Post]
      operationId: listUserPosts
      summary: 游标分页查询用户动态
      description: |
        固定按 publishedAt DESC、postId DESC 排序。匿名只返回 PUBLISHED + PUBLIC；
        携带登录令牌时实时校验关注关系或作者本人权限后，可补充 FOLLOWERS_ONLY 或 PRIVATE。
        作者账号不可用或资料未完成时统一返回 404；历史动态保留在数据库中但不公开分发。
      x-implementation-status: IMPLEMENTED
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FilterBoundCursor"
        - $ref: "#/components/parameters/PostLimit"
      responses:
        "200":
          description: 用户动态分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostPageResult"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/feed/following:
    get:
      tags: [Feed]
      operationId: listFollowingFeed
      summary: 游标分页查询关注流
      description: |
        返回当前用户所关注用户和企业的 PUBLISHED + PUBLIC/FOLLOWERS_ONLY 动态；
        查询时实时校验关注关系和作者公开可用状态；作者隐藏、停用或企业不可用时保留关注关系但不分发内容。
        固定按 publishedAt DESC、postId DESC 排序。
        不包含 PRIVATE、推荐算法、物化时间线或异步扇出。
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: "#/components/parameters/FilterBoundCursor"
        - $ref: "#/components/parameters/PostLimit"
      responses:
        "200":
          description: 关注流分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostPageResult"
        "401":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/feed/recommended:
    get:
      tags: [Feed]
      operationId: listRecommendedFeed
      summary: 游标分页查询推荐流
      description: |
        允许匿名访问。匿名用户、没有有效兴趣标签的登录用户和推荐依赖不可用时，
        返回公开动态的 publishedAt DESC、postId DESC 时间倒序结果。
        已登录且有有效兴趣标签时，使用 Ollama bge-m3 生成查询向量，
        从最近 90 天的公开动态中由 Qdrant 粗排，再由 Java 按语义、新鲜度、关注作者和同城规则精排。
        推荐结果只允许 PUBLISHED + PUBLIC 动态，绝不返回 FOLLOWERS_ONLY、PRIVATE、DRAFT、HIDDEN 或 DELETED；
        MySQL 在返回前重新校验动态、作者和公开可见性。Redis 快照保留 10 分钟，
        推荐游标独立绑定用户、快照、模式和首请求 limit；快照过期或游标无效返回 10001/400。
        本接口不改变 GET /api/v1/feed/following 的实时关注流语义。
      x-implementation-status: IMPLEMENTED
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RecommendedCursor"
        - $ref: "#/components/parameters/RecommendedLimit"
      responses:
        "200":
          description: 推荐动态分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPostPageResult"
        "400":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrganizationId:
      name: organizationId
      in: path
      required: true
      description: 企业业务编号；JavaScript 客户端必须按字符串保存。
      schema:
        $ref: "#/components/schemas/Id"
...
    UserId:
      name: userId
      in: path
      required: true
      description: 用户业务编号；JavaScript 客户端必须按字符串保存。
      schema:
        $ref: "#/components/schemas/Id"
    PostId:
      name: postId
      in: path
      required: true
      description: 动态业务编号；JavaScript 客户端必须按字符串保存。
      schema:
        $ref: "#/components/schemas/Id"
...
    FilterBoundCursor:
      name: cursor
      in: query
      required: false
      description: |
        M4 列表使用的后端不透明分页游标，客户端不得解析或改写。游标与产生它的筛选条件和排序方式绑定；
        改变筛选或排序后复用旧游标属于无效请求，返回 HTTP 400。
      schema:
        type: string
        minLength: 1
        maxLength: 512
    RecommendedCursor:
      name: cursor
      in: query
      required: false
      description: |
        推荐流专用的不透明游标，客户端不得解析或改写。游标绑定当前用户、推荐模式、
        Redis 快照和首请求 limit；快照有效期为 10 分钟。游标被篡改或快照过期返回 10001/400，
        游标由服务端使用 HMAC-SHA256 签名，客户端应从第一页重新请求；不得与公开动态、
        关注流或其他列表游标混用。
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
    PostLimit:
      name: limit
      in: query
      required: false
      description: 单页返回数量。
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    RecommendedLimit:
      name: limit
      in: query
      required: false
      description: 推荐流单页返回数量。
      schema:
        type: integer
        minimum: 1
        maximum: 20
        default: 20
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        创建命令的幂等键。结果在当前操作人、发布主体和键值范围内保留 24 小时。
        重放不延长窗口；活动或动态草稿删除也不缩短原 24 小时窗口。
        窗口到期后由请求事务先同步释放过期占用，再允许同键重新创建；后台清理延迟不得延长窗口。
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
  responses:
    VoidSuccess:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/VoidResult"
    Error:
      description: |
        统一错误响应。常见业务码：
        10000-10999 通用错误；11001-11005 登录会话；
        11101-11103 手机与设备；11201-11203 资料；
        11301-11302 协议；11401-11404 文件；11501-11503 账号数据；
        11601 幂等键冲突；11602 活动状态冲突；11603 乐观锁版本冲突；11604 活动容量冲突；
        11605 幂等请求对应的活动草稿已删除。
        11701 报名尚未开始；11702 报名已经截止；11703 活动名额已满；
        11704 活动发起者不能报名自己的活动；11705 活动已经开始；
        11706 活动已取消；11707 活动已结束；11708 当前参与记录不可由发起者移除；
        11709 已被活动发起者移除，不能再次报名；11710 不符合活动报名性别限制。
        11801 不能关注自己。
        11901 动态创建幂等键冲突；11902 动态状态冲突；11903 动态版本冲突；
        11904 幂等请求对应的动态草稿已删除；11905 动态内容未通过安全检查；
        11906 引用活动不可用。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResult"
...
  schemas:
    Id:
      type: string
      minLength: 1
      pattern: "^[1-9][0-9]*$"
      example: "1900000000000000201"
    Instant:
      type: string
      format: date-time
      example: "2026-07-29T08:00:00Z"
    ResultMeta:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          example: 成功
        requestId:
          type: string
          example: 019fa7f8b1aa7253
        timestamp:
          $ref: "#/components/schemas/Instant"
    ErrorResult:
      type: object
      required: [code, message, data, requestId, timestamp]
      properties:
        code:
          type: integer
          minimum: 1
          example: 10001
        message:
          type: string
          example: 参数校验失败
        data:
          nullable: true
          description: 参数校验失败时可能为 FieldValidationError 数组。
        requestId:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
    VoidResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              nullable: true
              example: null
...
    PublicActivityStatus:
      type: string
      enum: [PUBLISHED, CANCELLED, ENDED]
      default: PUBLISHED
...
    PostAuthorType:
      type: string
      enum: [USER, ORGANIZATION]
...
    ManagedPostStatus:
      type: string
      enum: [DRAFT, PUBLISHED, HIDDEN]
      description: 管理列表不返回已软删除动态。
    PostVisibility:
      type: string
      enum: [PUBLIC, FOLLOWERS_ONLY, PRIVATE]
      default: PUBLIC
      description: |
        PUBLIC 允许所有人读取；FOLLOWERS_ONLY 在读取时实时检查当前关注关系，个人作者本人或
        企业当前有效且唯一的 OWNER 也可读取自己主体的动态；
        PRIVATE 个人动态只允许作者本人，企业动态只允许当前有效且唯一的 OWNER。
...
    PostDraftCreateRequest:
      type: object
      additionalProperties: false
      description: |
        只保存 DRAFT。标题、正文、图片或活动引用至少存在一项；空白字符串按无内容处理。
        本期不接收 mentions、topicIds、location 或 declaration。
      properties:
        title:
          type: string
          maxLength: 20
          nullable: true
        content:
          type: string
          maxLength: 5000
          nullable: true
        visibility:
          $ref: "#/components/schemas/PostVisibility"
        activityId:
          description: 只接受当前公开可读的 PUBLISHED、CANCELLED 或 ENDED 活动；不可用时返回 11906/409。
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Id"
        mediaFileIds:
          type: array
          maxItems: 5
          uniqueItems: true
          default: []
          items:
            $ref: "#/components/schemas/Id"
    PostDraftReplaceRequest:
      type: object
      additionalProperties: false
      required: [version, visibility]
      description: 完整替换 DRAFT；标题、正文、图片或活动引用至少存在一项。
      properties:
        version:
          type: integer
          minimum: 0
        title:
          type: string
          maxLength: 20
          nullable: true
        content:
          type: string
          maxLength: 5000
          nullable: true
        visibility:
          $ref: "#/components/schemas/PostVisibility"
        activityId:
          description: 只接受当前公开可读的 PUBLISHED、CANCELLED 或 ENDED 活动；不可用时返回 11906/409。
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Id"
        mediaFileIds:
          type: array
          maxItems: 5
          uniqueItems: true
          default: []
          items:
            $ref: "#/components/schemas/Id"
    PostAuthorSummary:
      type: object
      required: [authorType, authorId, displayName, avatar]
      properties:
        authorType:
          $ref: "#/components/schemas/PostAuthorType"
        authorId:
          $ref: "#/components/schemas/Id"
        displayName:
          type: string
          minLength: 1
          maxLength: 256
        avatar:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PublicImageView"
    PostActivityCard:
      type: object
      description: 引用活动后来变为不可见时，动态仍可读取，但 activity 返回 null。
      required: [activityId, status, title, cover, startsAt, endsAt]
      properties:
        activityId:
          $ref: "#/components/schemas/Id"
        status:
          $ref: "#/components/schemas/PublicActivityStatus"
        title:
          type: string
          minLength: 1
          maxLength: 20
        cover:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PublicImageView"
        startsAt:
          $ref: "#/components/schemas/Instant"
        endsAt:
          $ref: "#/components/schemas/Instant"
    PublicPost:
      type: object
      required:
        - postId
        - author
        - visibility
        - title
        - content
        - media
        - activity
        - publishedAt
      properties:
        postId:
          $ref: "#/components/schemas/Id"
        author:
          $ref: "#/components/schemas/PostAuthorSummary"
        visibility:
          $ref: "#/components/schemas/PostVisibility"
        title:
          type: string
          minLength: 1
          maxLength: 20
        content:
          type: string
          maxLength: 5000
          nullable: true
        media:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/PublicImageView"
        activity:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PostActivityCard"
        publishedAt:
          $ref: "#/components/schemas/Instant"
    ManagedPostSummary:
      type: object
      required:
        - postId
        - author
        - status
        - visibility
        - title
        - content
        - media
        - activity
        - version
        - publishedAt
        - updatedAt
      properties:
        postId:
          $ref: "#/components/schemas/Id"
        author:
          $ref: "#/components/schemas/PostAuthorSummary"
        status:
          $ref: "#/components/schemas/ManagedPostStatus"
        visibility:
          $ref: "#/components/schemas/PostVisibility"
        title:
          type: string
          maxLength: 20
          nullable: true
        content:
          type: string
          maxLength: 5000
          nullable: true
        media:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/PublicImageView"
        activity:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PostActivityCard"
        version:
          type: integer
          minimum: 0
        publishedAt:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
    ManagedPostDetail:
      allOf:
        - $ref: "#/components/schemas/ManagedPostSummary"
        - type: object
          required: [createdAt, hiddenAt]
          properties:
            createdAt:
              $ref: "#/components/schemas/Instant"
            hiddenAt:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/Instant"
    PublicPostPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicPost"
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
    ManagedPostPage:
      type: object
      required: [items, nextCursor, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ManagedPostSummary"
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
    PublicPostPageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PublicPostPage"
    PublicPostResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/PublicPost"
    ManagedPostPageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManagedPostPage"
    ManagedPostDetailResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ManagedPostDetail"
    PublicImageView:
      type: object
      required: [fileId, url]
      properties:
        fileId:
          $ref: "#/components/schemas/Id"
        url:
          type: string
          example: /api/v1/files/1900000000000000301/content
...